-- Added support for initializing and mounting filesystems on block devices.
-- Added the ability to determine the filesystem size from the
size of the underlying block device.

-- Added logic to enforce limits on the overall filesystem
dimensions arising from limitations in the v6 code.

-- Added code to flush data to underling storage whenever the
filesystem is synced or unmounted.

-- Simplified option parsing code and improved associated
error reporting.

-- Changed the way errors are reported by the virtual disk
interface.

+++ ancient-src/v6/ken/alloc.c
@@ -146,11 +146,11 @@ int16_t
 badblock(struct filsys * afp, int16_t abn, int16_t dev)
 {
 	register struct filsys *fp;
-	register int16_t bn;
+	register uint16_t bn;
 
 	fp = afp;
-	bn = abn;
-	if (bn < fp->s_isize+2 || bn >= fp->s_fsize) {
+	bn = (uint16_t)abn;
+	if (bn < (uint16_t)fp->s_isize+2 || bn >= (uint16_t)fp->s_fsize) {
 		prdev("bad block", dev);
 		return(1);
 	}
